Solved speech recognition voice commands + more
Voice commands are now processed when repeated.
VoiceBite Icon has been changed to fill the entire square.
App reconfigured to be able to deploy on iOS 15.5 and up -> navigationStacks were changed to NavigationViews to accomodate this. Fonts were also changed as FontWeight is no longer available.
User preferences are now stored in firebase.